Bug#67156: Sporadic query cache related crash in pthread_rwlock_init()
Reinitializing the query cache might lead to a crash inside
pthread_rwlock_init() on Mac OS X. Mac OS X's pthread_rwlock_init()
implements a check to detect attempts to reinitialize a previously
initialized but not yet destroyed read/write lock. This check works by
looking for a signature within the rwlock object and, if the signature
matches, some pointer variables within the object are dereferenced.

When MySQL's query cache is resized (e.g. SET query_cache_size = ...),
or simply deinitialized, the memory used to allocate blocks is freed
without explicitly deinitializing the allocated blocks. In particular,
'queries blocks' contain rwlock objects that are not properly destroyed.

If the query cache is resized, and the same memory region is used, and
a new query block is allocated in the same position, attempting to
initialize the block's read/write lock might crash if the read-write
lock object signature was preserved but yet somehow some of its fields
were corrupted.

The solution is to properly destroy the used read-write lock objects.