Increase time resolution of expire_logs_days
In some scenarios, it is necessary to purge the binary logs more
frequently than once a day. In these cases, @@expire_logs_days cannot
be used for automatic binary log file removal because of its whole-day
resolution.

In order to allow for more fine-grained control over the automatic
removal period, the type of the @@expire_logs_days variable is changed
to decimal where the fractional part represents the fraction of one
day. For example, 12 hours can be expressed as 0.5 day.