Thread state is "query end" while blocked writing to binary log
When binary logging is enable, the thread state is set to "query end"
most recently before writing to the binary log, so any time taken in
that (due to contention or I/O) shows up as "query end" which makes it
difficult to determine what is the current state of the thread.

The solution is to set more appropriate thread states in any binary
log related activities that may take time. For example, now whenever
an event is being writing to the binary log, the thread state is set
to "Writing an event to the binary log".