Bug #14226171  EXCESSIVE ROW LOCKING WITH UPDATE IN 5.5.25
When a DML statement is issued, and if the index merge 
access method is chosen, then many rows from the
storage engine will be locked because of the way the
algorithm works.  Many rows will be locked, but they
will not be part of the final result set.

To reduce the excessive locking, the locks of unmatched
rows are released by this patch.  This patch will 
affect only transactions with isolation level
equal to or less stricter than READ COMMITTED. This is
because of the behaviour of ha_innobase::unlock_row().

rb://1296 approved by jorgen and olav.

--BZR--
revision-id: annamalai.gurusami@oracle.com-20121019111348-2989cc8r03cp3ftw
property-audit-revid: annamalai.gurusami@oracle.com-20121019111348-xcbsqfqu5u4e9b68
property-branch-nick: mysql-5.5
testament3-sha1: 09a70cf6ab2bbd95bce5c5631f42fdac0fc8f69e