Backport from 5.6: Bug#14769820 ASSERT FLEN == LEN IN ALTER TABLE ... ADD UNIQUE KEY
A bogus debug assertion failure occurred when reporting a duplicate
key on a column prefix of a CHAR column.

This is a regression from Bug#14729221 IN-PLACE ALTER TABLE REPORTS ''
INSTEAD OF REAL DUPLICATE VALUE FOR PREFIX KEYS. The assertion is only
present when UNIV_DEBUG is defined (which it is in debug builds
starting from MySQL 5.5). It is a case of overasserting.

Fix approved by Inaam Rana on IM.
--BZR--
revision-id: marko.makela@oracle.com-20121022191033-kdie8oejzjto9a17
property-audit-revid: marko.makela@oracle.com-20121022191033-7bsypio82zg0czk9
property-branch-nick: mysql-5.1
testament3-sha1: fc09f64387a77bc7d695a994ec44f72c8877864c