Bug#67433: Using SET GLOBAL SQL_LOG_BIN should not be allowed
Earlier in MySQL 5.5 development cycle, the SQL_LOG_BIN variable was
made to be both global and session-scoped, instead of only session as
it was in previous releases. The problem is that usage of SQL_LOG_BIN
at the global scope can be quite dangerous.

When SQL_LOG_BIN is changed globally, it does not immediately take
effect for any sessions. It takes effect by becoming the default value
for new sessions, and this setting is kept and cached for the duration
of that session (unless changed). Setting it intentionally is unlikely
to have a useful effect under any circumstance; setting it
unintentionally, such as while intending to use SET [SESSION] is
potentially disastrous.

Accidentally using SET GLOBAL SQL_LOG_BIN will not show an immediate
effect to the user, it won't have the desired session-level effect,
and might cause other potential problems with local-only maintenance
being logged and executed on slaves. Furthermore, a bigger problem is
with the long term and possibly permanent damage which may result in
irrecoverable or difficult data loss with binary logging being
disabled for new (and unrelated) sessions.

The solution is to revert the change and again make SQL_LOG_BIN a
session-only variable, generating an error if it is used with SET
GLOBAL.