A non-blocking mode for ALTER TABLE.
While an ALTER TABLE is executing, the table whose definition is being
changed can usually be read, but not updated, by other sessions. Such
writes are blocked until the ALTER TABLE operation completes or a lock
wait timeout occurs.

In a sharded system, it would be better if conflicting operations were
aborted rather than blocked, thereby eliminating delays due to lock
waits and allowing them to be redirect to another shard.

In order to allow for this behavior, a new ALTER TABLE option named
NO_WAIT is introduced. When NO_WAIT is set, conflicting locks requests
that would normally block are aborted instead.