Bug#67476: Innodb_buffer_pool_read_ahead_evicted is inaccurate
If a page being read into the buffer pool is made "young" (moved
to the head of the LRU), its time of first access it not properly
recorded (that is, it remains with a value of zero). When the page
eventually gets evicted, the page is counted as a read-ahead page
that was evicted without having been accessed by queries.

The solution is to ensure that the first access time is recorded
even if the page is moved to the head of the LRU.