ALTER TABLE in exclusive mode.
In certain cases, it is desirable to control the lock mode to be used
for an ALTER TABLE operation, especially in a scenario where it is
appropriate to restrict access to the table while the operation is in
progress. For example, an exclusive lock mode could be used in
conjunction with NO_WAIT to abort any attempts to use a table while
its structure is being changed.

In order to support such scenarios, this changes adds a LOCK clause
for ALTER TABLE. Inspired by the LOCK clause introduced in MySQL 5.6,
this clause can be used to enforce a specific lock mode for an alter
table operation. The only supported lock mode parameter is EXCLUSIVE,
which will block any attempts to read or write to the table while the
operation is in progress.