Make RPMs for ULN build again.
A change to "vio/viosslfactories.c" in August, 2012,
broke a patch which is to be applied during the build
of ULN RPMs.
The patch file is
"packaging/rpm-uln/mysql-chain-certs.patch"

This change bypasses the problem by not trying to apply
the patch.

This is a regression and must be fixed, not bypassed.
--BZR--
revision-id: joerg.bruehe@oracle.com-20121107193254-v51thnf4wc4xmldo
property-audit-revid: joerg.bruehe@oracle.com-20121107193254-f7zodfhdlyxjkkkp
property-branch-nick: place-uln-5.5
testament3-sha1: 91335aa40243e54f2edde2a6dca036e5dd919ef0