Expose the level of each B-tree page in the buffer pool
Although the INNODB_BUFFER_PAGE table shows the number of records
within a page, it is not clear whether these records are row data or
internal b-tree pointers. For this reason, a new LEVEL column is
introduced which shows the level of a page in the B-tree.

Additionally, the direction of insertion encoded in InnoDB data
pages is now available in the DIRECTION column, and the number of
consecutive inserts in that direction is available in the N_DIRECTION
column. These columns are useful to determine whether records are
being inserted in an ascending or descending sequence.