Expose the segment fill factor as a configurable setting.
When allocating a page from a segment, InnoDB has an internal fill
factor setting that is used to decide whether to allocate a new extent
or to use an unused page. In short, the segment fill factor is used to
determine how full to make the segment before extending it. The fill
factor resembles a ratio of the number of unused pages to the total
number of pages (1:8), meaning that some unused pages are left behind
for future (contiguous) expansion of the index. If taking a page would
cause the ratio to decrease, a new extent is allocated instead.

A problem with this default fill factor is that it may lead to a lot of
wasted space if the segment becomes fragmented. The larger the table,
the more wasted space there might be. Depending on the performance
characteristics of the hard disk, it might make more sense to increase
the fill factor if, for example, seek time is not a concern.

This change introduces a new variable named innodb_segment_fill_factor
that can be used to control the fill factor. Its value specifies the
percentage of space on a segment to be used, reserving the remainder as
free space for future growth. The default value is set to 88%.