Expose the index fill factor as a configurable setting.
When extending a clustered index to the left or right, InnoDB attempts
to fill leaf pages only up to a certain percentage, at which point the
page is split. The remaining space is reserved for updating rows
belonging to the page, attempting to make it likely that an updated row
will be stored on the same page. This percentage is also know as the
index fill factor and is used to determine how much to fill leaf pages
and how much space to reserve for future expansion, with the aim of
reducing the need to split pages as the data grows. Currently, InnoDB
reserves 1/16 (6.25%) of the page for future updates of records.

A problem with this imposed fill factor is that for tables whose entries
are never updated, or whose records are of fixed size (meaning that the
record can be update in place), the reserved space might lead to
premature page splits and also cause a lot of space to be wasted. In any
case, if the pattern of insertion/modification can be determined, it
might be worthwhile to tweak the fill factor to reduce or increase the
amount of space that is reserved.

This change introduces a new variable name innodb_index_fill_factor that
can be used to control the fill factor for clustered indexes. Its value
specifies the percentage of space on a leaf page to be used, reserving
the remainder as free space. The default value is set to 94%.