Bug#13545447 RPL_ROTATE_LOGS FAILS DUE TO CONCURRENCY ISSUES IN REP. CODE
RPL_ROTATE_LOGS has been failing sporadically in what seems a
problem related to routines that update the coordinates. However,
the test lacks proper assert statments and because of this the
debug information upon failure simply points to the content
mismatch between the test and the result file.

Not as a solution, but as a improvement to the test to better
debug this failure, new assert statments were added to the test.

@rpl_rotate_logs.test
  Added new assert statments reducing the
  dependency on the result file.

@rpl_rotate_logs.result
  Added new content to the result file to
  match the test changes
--BZR--
revision-id: pedro.gomes@oracle.com-20121204160948-5wtp9jlbfsg720br
property-audit-revid: pedro.gomes@oracle.com-20121204160948-ximr3v8pdsztyfro
property-branch-nick: mysql-5.5
testament3-sha1: 0c96c84db943ec1628f03d463c69bbea4e9895a8