Bug#16060864 SEGMENTATION FAULT IN PERFORMANCE_SCHEMA WITH HISTORY SIZE 0
Before this fix, configuring the server with:
- performance_schema_events_waits_history_size=0
- performance_schema_events_waits_history_long_size=0
could cause a crash in the performance schema.

These settings to 0 are intended to be valid and supported,
and are in fact working properly in mysql 5.6 and up already.

This fix backports the code fix and test cases from mysql 5.6
to the mysql 5.5 release.

--BZR--
revision-id: marc.alff@oracle.com-20130102100055-rpand3ksmd0ywaz6
property-audit-revid: marc.alff@oracle.com-20130102100055-gpf3ku1z55epyvug
property-branch-nick: mysql-5.5-bug16060864
testament3-sha1: ec44d45cd05500c95891675cdec3d57c6b22e4eb