The fractional part of the def, min and max values of system variables is ignored.
Allow the default, minimum and maximum values of system variables of
type double to have a meaningful fractional part. Since the command-line
option parsing interface (my_getopt) uses fields of type unsigned long
long (ull) to store these values, the double values were being stored in
a lossy way that discards the fractional part.

This change introduces a couple of functions that can be used to store
the raw representation of a double value in the raw bits of unsigned
long long field in a way that the binary representation remains the
same. Using these function the real value can be converted back and
forth between the types.