Change the type of the system variables innodb_segment_fill_factor and innodb_index_fill_factor to double.
These variables are used to represent percentages but could not be set
to percentages with fractional parts. Furthermore, their initial default
values were fractional percentages, but had to be changed because the
interface did not support them. Now these system variables are made of
type double and their default values are restored to their original
form.