InnoDB reserves an excessive amount of space in large tables for write operations.
When performing operations that are expected to expand a table (for
example, allocate new pages due to a page split), InnoDB currently
preallocates and reserves up to 1% of the total size of the tablespace
as a measure to ensure that enough free extents (that is, disk space)
are available for the operation and to ensure that if running out of
disk space, these operations are preemptively failed as to reserve any
remaining free space to operations that end up freeing space (that is,
delete data).

The percentage is reasonable for tables smaller than a few gigabytes,
but not for tables sized at tens of gigabytes or more, at which point
the percentage won't correctly estimate the free space needed to perform
operations and may cause an excessive amount of free extents to be
preallocated. Also, this reservation approach is of dubious need
considering that disk space usage is normally monitored and that
transactions normally won't cause large expansions. In the worst case,
it is even possible to expand the underlying filesystem if the database
goes into a deadlock where it lacks free space for cleaning operations.

The solution is to provide a way to either completely disable free
extents reservation or to control the amount of free extents that are
reserved for such operations. This changes introduces two new system
variables to accomplish both. The variable innodb_reserve_free_extents
can be used to enable or disable free extents reservation and
innodb_free_extents_reservation_factor can be used to control what
percentage of a space size is reserved for operations that may cause
more space to be used.