Relocate directory "SPECIFIC-ULN/ to "packaging/rpm-uln/".
Revision joerg.bruehe@oracle.com-20121108180644-n67cgi9o3lewt99g, which
renames the directory "SPECIFIC-ULN/" to "packaging/rpm-uln/" broke the
importation of changes from the upstream MySQL bazaar tree to git.

The cause of the problem is that the bzr-git plugin, which is used to
import said changes, does not properly handle directories being renamed
(see Launchpad Bug#818318). In this case, the import succeeds but ends
up creating a broken tree.

In order to work around this problem, the "SPECIFIC-ULN/" directory is
removed and re-added (with same file ids) as "packaging/rpm-uln/" in a
later revision. This basically mimics the rename as  a remove and add.
This way, the bzr-git plugin is able to correctly handle the directory
move.

From this point on, upstream changes will need to be merged, but this
shouldn't cause any conflicts.

--BZR--
revision-id: davi@twitter.com-20130105031502-4nn534ijots9xojy
property-branch-nick: mysql-5.5-rewrite
testament3-sha1: cecd2384199a68f2716bb9fed81df52f0913d996