Bug#68051: Killing a query inside InnoDB causes it to eventually crash with an assertion
Killing a query that is inside InnoDB searching for a row (specifically
in row_search_for_mysql) causes InnoDB to crash with an assertion failure
the next time the same table/cursor instance is used again.

This is a regression introduced by the fix for Bug#14704286 (revision
id: krunal.bauskar@oracle.com-20121015041950-v9r57913ucuh4f21, git
commit: 66c9023cea344b825349b7888de70b687bac1a4f), which introduces a
check in row_search_for_mysql to interrupt the function if the query has
been killed.

The problem is that the error code path in row_search_for_mysql used to
handle an interruption does not properly store the cursor position upon
exit, which might be required in some cases. This leads to an assertion
failure in btr_pcur_restore_position_func (btr0pcur.c:250) once the
cursor is used again.

The solution is to simply revert the original fix. The risks associated
with introducing an interruption point in a code path that was previously
uninterruptible far outweigh any benefits.