Expose operations counters per table.
Although MySQL has counters for the number of requests/operations
on tables (e.g. Handler_write), it is not possible to break down
the numbers per table, which would be useful in identifying what
types of operations are being performed on specific tables.

This change introduces a set of per table statistics counters for
handler operations. The counters are incremented for each specific
operation on a handler instance and once the handler instance is
closed, the counters are accrued in the table definition stored in
the definition cache. If the table definition is altered or flushed,
the counters are lost.

An information schema table named TABLE_STATISTICS is introduced to
provide the statistics counters for each table in the definition cache.
Each counter is displayed in a separate column that, for example,
records the number of insert, delete, or update operations, as well as
other operations, plus columns for the table and database names.