Revert MySQL revision f8a5acc (and any subsequent related changes).
The fix for Bug#14548159 (Numerous cases of incorrect identifier quoting
in replication) introduced a regression that breaks statement-based
replication by generating statements with malformed quoting which will
cause a slave to stop since such statements cannot be parsed.

For now, the solution is to simply revert the introduced changes.
Although the change was introduced to fix a security issue, it is more
important to ensure that new problems are not introduced, especially
since the security issue is low risk.