Expose total per-table space file size as Max_data_length.
Althought SHOW TABLE STATUS exposes, for InnoDB tables, the size of the
clustered and secondary indexes and the amount of free space, it does
not expose, nor it is possible to calculate, the total size of the
tablespace to which the table belongs — because any internal overhead is
deducted from the exposed sizes.

This change repurposes the Max_data_length field of SHOW TABLE STATUS to
expose the actual total size of the tablespace to which an InnoDB table
belongs to. Previously, this field was left unused (zeroed) for such
tables.