Bug#16208709 - CRASH IN GET_SEL_ARG_FOR_KEYPART ON SELECT DISTINCT 	       ON COL WITH COMPOSITE INDEX
This problem is caused by the patch for the bug#11751794.
While checking for the keypart covering non grouping attribute. we are not 
checking whether the root node of the SEL_ARG* tree for the index have any 
cvalue or not.
--BZR--
revision-id: neeraj.x.bisht@oracle.com-20130129043500-kt51kec5pftksil2
property-audit-revid: neeraj.x.bisht@oracle.com-20130129043500-f93e4b3wcvbin1xc
property-branch-nick: 5.1
testament3-sha1: 65cb247d0972dcb3c9ac6676e7047e7f21da2f54