BUG#16207679: MISSING ERROR WHEN RESIGNAL TO MYSQL_ERROR=5
Analysis:
--------

As part of the fix for Bug#11757464, the 'out of memory' error
condition was not pushed to the diagnostic area as it requires
memory allocation. However in cases of SIGNAL/RESIGNAL 'out of
memory' error, the server may not be out of memory. Hence it
would be good to report the error in such cases.

Fix:
---

Push only non fatal 'out of memory' errors to the diagnostic area.
Since SIGNAL/RESIGNAL of 'out of memory' error may not be fatal,
the error is reported.
--BZR--
revision-id: nisha.gopalakrishnan@oracle.com-20130201041927-e79fbqqkopj1vfbf
property-audit-revid: nisha.gopalakrishnan@oracle.com-20130201041927-22546ajxshk7vkow
property-branch-nick: mysql-5.5-16207679
testament3-sha1: 257d0a4a73cbdbdb612ac66756858de1e34242d9